float format on price property in space model

+++ app/models/space.rb
@@ -10,5 +10,5 @@ has n, :calendars, through: Resource
   property :name, String
   property :address, Text
   property :postcode, Text
-  property :price, Decimal
+  property :price, Float
 end
